YAY-114 refactor: transfer  PostCreatorSlider и modals from entities to features
