+++ src/entities/post/ui/Post/PostClientWrapper/PostClientWrapper.tsx
@@ -2,7 +2,7 @@
 import type { PostByIdType } from '@/features/post/api'
 import { Post } from '@/entities/post/ui/Post'
 import { usePostController } from '@/features/post/hooks'
-import { EditPostModal } from '@/entities/post/ui/modals/EditPostModal'
+import { EditPostModal } from '@/features/post-creator/ui/modals/EditPostModal'
 
 export function PostClientWrapper({ post }: { post: PostByIdType }) {
    const { closePostModal, closeEditPostModal, isEditOpen } = usePostController()
