+++ src/entities/user/model/useUserProfile.ts
@@ -1,3 +1,4 @@
+'use client'
 import { useParams } from 'next/navigation'
 import { useGetUserByLoginQuery } from '../api/user.service'
