+++ src/features/post-creator/ui/modals/CroppingModal/CroppingModal.stories.tsx
