+++ src/features/post-creator/ui/modals/FilterModal/FilterBlock/FiltersBlock.tsx
