+++ src/features/post-creator/ui/modals/FilterModal/FilterBlock/assets/Filter.jpg
